requirements.txt
asgiref==3.5.2
Django==4.1.4
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Pillow==9.3.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.2.0
